pyproject.toml
[tool.poetry]
name = "daily-ops"
version = "1.0.0"
description = ""
authors = ["Alistair Stark <astark@spiria.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
SQLAlchemy = "^1.4.29"
alembic = "^1.7.5"
psycopg2 = "^2.9.3"
black = "^22.3.0"
isort = "^5.10.1"
pytest-asyncio = "^0.16.0"
asyncio = "^3.4.3"
asyncpg = "^0.25.0"
httpx = "^0.21.3"
cryptography = "^36.0.1"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
pytest-freezegun = "^0.4.2"
bs4 = "^0.0.1"
pytz = "^2021.3"
lxml = "^4.7.1"
bcrypt = "^3.2.2"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
